DSET-3995 refactor - rename os_util to be according go conventions
