+++ internal/pkg/os/util/util_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package os_util
+package util
 
 import (
 	"fmt"
